backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
pyjwt
python-multipart
aiofiles
